fixed problem with null transient. was getting a negative velocity.  this messed everything up.  Should check for physical inconsistencies

+++ PrecursorSolve/src/inter_shape_fcns.f90
@@ -11,9 +11,9 @@ subroutine inter_shape_fcns (xi, h)
 
     implicit none
 !---Dummy variable
-    real :: xi 
+    double precision :: xi 
     !real, dimension(3) :: elem_coord
-    real :: h
+    double precision :: h
 
 !---local
     integer :: i
